pyproject.toml
[project]
name = "api-postgres"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.115.14",
    "passlib[bcrpyt,bcrypt]>=1.7.4",
    "psycopg[binary]>=3.2.9",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
